ui & apertad: implement semi file verification
Now implemented size 0 checks on libs, and re downloading on error.
Also, implemented basic checks for assets existence as well as forge
version existing.

Also implemented better UX for loading bar, as well as fully exiting on
close and not hanging around on the back.

This version continues the breakage of Linux, for unknown reasons.